+++ src/loaders/index.ts
@@ -1,4 +1,5 @@
 export * from './services.loader';
+export * from './custom-api.loader';
 export * from './admin-api.loader';
 export * from './store-api.loader';
 export * from './repository.loader';
@@ -8,6 +9,7 @@ export * from './plugins.loader';
 export * from './migrations.loader';
 export * from './validators.loader';
 export * from './modules.loader';
+export * from './subscribers.loader';
 
 export * from '../core/types';
 export * from '../core/utils';