+++ src/modules/multi-tenancy/tenant.module.ts
@@ -1,11 +1,15 @@
 import { TenantMiddleware } from './tenant.middleware';
-import { Tenant } from './tenant.entity';
-import TenantRepository from './tenant.repository';
-import TenantMigration1647204670085 from './1647204670085-tenant.migration';
 import { Module } from '../../decorators';
 import { TenantService } from './tenant.service';
+import { MedusaDynamicModule, ModuleInjectionOptions } from '../../core';
+import { ConfigModule } from './types';
 
-@Module({
-	imports: [TenantMigration1647204670085, TenantMiddleware, Tenant, TenantRepository, TenantService],
-})
-export class TenantModule {}
+@Module()
+export class TenantModule implements MedusaDynamicModule {
+	async forRoot(configModule: ConfigModule): Promise<ModuleInjectionOptions> {
+		if (configModule.multi_tenancy?.enable) {
+			return { imports: [TenantMiddleware, TenantService] };
+		}
+		return { imports: [] };
+	}
+}